[Refactor: Section] 패키지 분리
