+++ src/main/java/subway/line/LineResponse.java
@@ -2,6 +2,7 @@ package subway.line;
 
 import lombok.*;
 import subway.*;
+import subway.line.section.*;
 
 import java.util.*;
 import java.util.stream.*;
