+++ src/main/java/subway/line/LineService.java
@@ -4,6 +4,7 @@ import lombok.*;
 import org.springframework.stereotype.*;
 import org.springframework.transaction.annotation.*;
 import subway.*;
+import subway.line.section.*;
 
 import java.util.*;
 import java.util.stream.*;
