+++ src/main/java/subway/line/section/DuplicateSectionStationException.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.section;
 
 public class DuplicateSectionStationException extends RuntimeException{
