+++ src/main/java/subway/line/section/NotLastDownStationException.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.section;
 
 public class NotLastDownStationException extends RuntimeException {
