+++ src/main/java/subway/line/section/SectionAddRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.section;
 
 import lombok.*;
