+++ src/main/java/subway/line/section/SectionService.java
@@ -1,9 +1,10 @@
-package subway.line;
+package subway.line.section;
 
 import lombok.*;
 import org.springframework.stereotype.*;
 import org.springframework.transaction.annotation.*;
 import subway.*;
+import subway.line.*;
 
 @Service
 @Transactional
